refactor create modals
