+++ js/data.js
@@ -1,12 +1,12 @@
 import {getData} from './api.js';
-import {showAlertMessage} from './user-form.js';
+import {showAlertMessage} from './modals.js';
 
 const NUMBER_OF_PHOTOS = 25;
 
 let dataList = [];
 
 export const initData = (cbSuccess) => {
-  getData (
+  getData(
     (photos) => {
       dataList = photos.slice(0, NUMBER_OF_PHOTOS);
       if (cbSuccess) {cbSuccess(dataList);}
