+++ js/modals.js
@@ -1,7 +1,11 @@
 const errorTemplate = document.querySelector('#error').content.querySelector('.error');
 const successTemplate = document.querySelector('#success').content.querySelector('.success');
+const imgUploadOverlay = document.querySelector('.img-upload__overlay');
+
 const successFragment = document.createDocumentFragment();
 const errorFragment = document.createDocumentFragment();
+
+const ALERT_SHOW_TIME = 5000;
 
 const triggerOnEsc = (evt, element) => {
   if (evt.key === 'Escape') {
@@ -48,4 +52,41 @@ export const showSuccessMessage = () => {
   document.addEventListener('keydown', (keyEvt) => triggerOnEsc(keyEvt, sectionSuccess));
 };
 
+export const showAlertMessage = (message) => {
+  const alertContainer = document.createElement('div');
+  alertContainer.classList.add('error-message');
+  alertContainer.textContent = message;
+  document.body.append(alertContainer);
 
+  setTimeout(() => {
+    alertContainer.remove();
+  }, ALERT_SHOW_TIME);
+};
+
+export const showUploadPopup = (cb, isReadyForClose, cbClose) => {
+  imgUploadOverlay.classList.remove('hidden');
+  document.body.classList.add('modal-open');
+  document.addEventListener('keydown', (evt) => keyDownHandler(evt, isReadyForClose, cbClose));
+  if (cb) {
+    cb();
+  }
+};
+
+export const closeUploadPopup = (cb, isReadyForClose) => {
+  imgUploadOverlay.classList.add('hidden');
+  document.body.classList.remove('modal-open');
+  document.removeEventListener('keydown', (evt) => keyDownHandler(evt, isReadyForClose, cb));
+  if (cb) {
+    cb();
+  }
+};
+
+//Функция объявлена декларативно, чтобы могла быть вызвана раньше, чем она объявлена
+function keyDownHandler (evt, isReadyForClose, cbClose) {
+  if (evt.key === 'Escape' && isReadyForClose) {
+    evt.preventDefault();
+    if (cbClose) {
+      closeUploadPopup(cbClose, isReadyForClose);
+    }
+  }
+}
